# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Detection event class. Seismic classes come from the STA/LTA detector; wifi_presence comes from the WiFi radar (CSI) motion detector.
 */
export type EventClass = typeof EventClass[keyof typeof EventClass];


export const EventClass = {
  unknown: 'unknown',
  footstep: 'footstep',
  vehicle: 'vehicle',
  multiple: 'multiple',
  wifi_presence: 'wifi_presence',
} as const;

/**
 * Which physical sensing channel produced the event.
 */
export type SensingChannel = typeof SensingChannel[keyof typeof SensingChannel];


export const SensingChannel = {
  seismic: 'seismic',
  rf: 'rf',
} as const;

export type GatewayNodeColor = typeof GatewayNodeColor[keyof typeof GatewayNodeColor];


export const GatewayNodeColor = {
  green: 'green',
  yellow: 'yellow',
  red: 'red',
} as const;

export interface GatewayNode {
  nodeId: number;
  /** @nullable */
  name?: string | null;
  color: GatewayNodeColor;
  reasons: string[];
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lon?: number | null;
  /** @nullable */
  lastSeen?: number | null;
  /** @nullable */
  batteryMv?: number | null;
  /** @nullable */
  noiseFloor?: number | null;
  /** WiFi radar (CSI) sensing active on this node */
  csiOn: boolean;
  /** CSI detector still calibrating its baseline */
  csiCalibrating: boolean;
  /**
     * CSI quiescent noise metric (baseline moving variance)
     * @nullable
     */
  csiNoise?: number | null;
}

export interface GatewayNodeList {
  nodes: GatewayNode[];
}

export interface GatewayAlert {
  nodeId: number;
  receivedAt: number;
  eventClass: EventClass;
  channel: SensingChannel;
  confidence: number;
  /** Seismic classes: peak band amplitude. wifi_presence: CSI motion metric x100. */
  peakAmp: number;
  /** @nullable */
  batteryMv?: number | null;
}

export interface GatewayAlertList {
  alerts: GatewayAlert[];
}

export interface GatewayError {
  error: string;
}

export type ListGatewayAlertsParams = {
/**
 * @minimum 1
 * @maximum 500
 */
limit?: number;
/**
 * Filter to one event class
 */
eventClass?: EventClass;
};