# lib/api-zod/src/generated/types/eventClass.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Detection event class. Seismic classes come from the STA/LTA detector; wifi_presence comes from the WiFi radar (CSI) motion detector.
 */
export type EventClass = typeof EventClass[keyof typeof EventClass];


export const EventClass = {
  unknown: 'unknown',
  footstep: 'footstep',
  vehicle: 'vehicle',
  multiple: 'multiple',
  wifi_presence: 'wifi_presence',
} as const;