# lib/api-zod/src/generated/types/gatewayAlert.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EventClass } from './eventClass';
import type { SensingChannel } from './sensingChannel';

export interface GatewayAlert {
  nodeId: number;
  receivedAt: number;
  eventClass: EventClass;
  channel: SensingChannel;
  confidence: number;
  /** Seismic classes: peak band amplitude. wifi_presence: CSI motion metric x100. */
  peakAmp: number;
  /** @nullable */
  batteryMv?: number | null;
}