# lib/api-zod/src/generated/types/gatewayAlertList.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GatewayAlert } from './gatewayAlert';

export interface GatewayAlertList {
  alerts: GatewayAlert[];
}