# lib/api-zod/src/generated/types/gatewayNode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GatewayNodeColor } from './gatewayNodeColor';

export interface GatewayNode {
  nodeId: number;
  /** @nullable */
  name?: string | null;
  color: GatewayNodeColor;
  reasons: string[];
  /** @nullable */
  lat?: number | null;
  /** @nullable */
  lon?: number | null;
  /** @nullable */
  lastSeen?: number | null;
  /** @nullable */
  batteryMv?: number | null;
  /** @nullable */
  noiseFloor?: number | null;
  /** WiFi radar (CSI) sensing active on this node */
  csiOn: boolean;
  /** CSI detector still calibrating its baseline */
  csiCalibrating: boolean;
  /**
     * CSI quiescent noise metric (baseline moving variance)
     * @nullable
     */
  csiNoise?: number | null;
}