# lib/api-zod/src/generated/types/sensingChannel.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Which physical sensing channel produced the event.
 */
export type SensingChannel = typeof SensingChannel[keyof typeof SensingChannel];


export const SensingChannel = {
  seismic: 'seismic',
  rf: 'rf',
} as const;